ES-Lint check
Description:
Disabled the Lint check for vue config file as there are some unused variables that are called upon. Need to disable for this as it will never go through the Lint check unless its changed.

Type of change:
- [x] Bug fix (non-breaking change)
- [ ] New feature (non-breaking change)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

Testing:
Tested in a separate repo.

Additional Notes:
nyaa nyaa nyaa