Improve deletion of previous user profile image
This commit refactors the `upload_image` function in `views.py` to improve the deletion of the previous user profile image. Previously, the code would always attempt to delete the previous image, even if it was the default profile picture. This resulted in unnecessary API calls to the S3 bucket.

Now, the code first checks if the user's profile picture is the default image. If it is, the deletion process is skipped. However, if the profile picture is a custom image, the previous image URL is retrieved and the corresponding image file is deleted from the S3 bucket.

This change improves the efficiency of the image deletion process and reduces unnecessary API calls.